+++ example/android/app/src/main/kotlin/com/feedmatter/flutter/sdk/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.feedmatter.flutter.sdk.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()